qrtr-control-socket: fix warnings with -Wsign-compare
  qmi-qrtr-control-socket.c: In function ‘qrtr_ctrl_message_cb’:
  qmi-qrtr-control-socket.c:216:24: error: comparison of integer expressions of different signedness: ‘gssize’ {aka ‘long int’} and ‘long unsigned int’ [-Werror=sign-compare]
    216 |     if (bytes_received < sizeof (ctrl_packet)) {
        |                        ^

+++ src/libqmi-glib/qmi-qrtr-control-socket.c
@@ -213,7 +213,7 @@ qrtr_ctrl_message_cb (GSocket           *gsocket,
         return FALSE;
     }
 
-    if (bytes_received < sizeof (ctrl_packet)) {
+    if ((gsize)bytes_received < sizeof (ctrl_packet)) {
         g_debug ("[qrtr] short packet received: ignoring");
         return TRUE;
     }
